Force vertex initializers to wrap every 360 degrees.
This isn't normally necessary, since any mathematically accurate implementation of the sine and cosine functions will repeat the functional pattern on a period of every 360 degrees, just as JavaScript's implementation does already.

It does however help show that the fifth vertex, 72*4 + 90, does exceed 360.  Instead of taking the sine of 378 degrees, we need only take the sine of 18 degrees.